[drop] Drop support for the deprecated Kuzzle Proxy (WIP)
