+++ lib/api/controllers/routerController.js
@@ -40,7 +40,7 @@ class RouterController {
   }
 
   /**
-   * Declares a new connection on a given protocol. Called by proxyBroker.
+   * Declares a new connection attahed to a network protocol.
    *
    * @param {RequestContext} requestContext
    */
@@ -58,7 +58,7 @@ class RouterController {
   }
 
   /**
-   * Called by proxyBroker: removes a connection from the connection pool.
+   * Removes a connection from the connection pool.
    *
    * @param {RequestContext} requestContext
    */
@@ -95,7 +95,6 @@ class RouterController {
     // Check only defined connection identifiers (some protocols might
     // not have one)
     return requestContext.connection.id === null
-      || requestContext.connection.misc.proxy === true
       || this.connections.has(requestContext.connection.id);
   }
