+++ lib/api/core/entrypoints/context.js
@@ -23,7 +23,7 @@
 
 const
   ClientConnection = require('./clientConnection'),
-  debug = require('../../../../kuzzleDebug')('kuzzle:entry-point:protocols'),
+  debug = require('../../../kuzzleDebug')('kuzzle:entry-point:protocols'),
   ProtocolBase = require('./protocols/protocol'),
   {
     errors,
