+++ lib/api/core/entrypoints/manifest.js
@@ -20,43 +20,30 @@
  */
 
 const
-  Bluebird = require('bluebird');
+  fs = require('fs'),
+  AbstractManifest = require('../abstractManifest');
 
-class EntryPoint {
-  constructor (kuzzle) {
-    this.kuzzle = kuzzle;
+class Manifest extends AbstractManifest {
+  constructor(kuzzle, pluginPath, protocol) {
+    super(kuzzle, pluginPath);
+    this.protocol = protocol;
   }
 
-  dispatch (event, data) {
-    // do nothing by default
-    return {event, data};
-  }
+  load() {
+    // @deprecated - there will be no fallback if no manifest.json file can be found
+    try {
+      fs.accessSync(this.manifestPath, fs.constants.R_OK);
+    } catch (e) {
+      if (this.protocol.protocol) {
+        this.kuzzle.log.warn(`[${this.path}] Protocols without a manifest.json file are deprecated.`);
+        this.name = this.protocol.protocol;
+        return;
+      }
+      // else let the parent class throw its own exception
+    }
 
-  /**
-   * @returns {Promise}
-   */
-  init () {
-    return Bluebird.resolve();
+    super.load();
   }
-
-  /**
-   * @param {string} channel
-   * @param {string} connectionId
-   */
-  joinChannel (channel, connectionId) {
-    // do nothing by default
-    return {channel, connectionId};
-  }
-
-  /**
-   * @param channel
-   * @param connectionId
-   */
-  leaveChannel (channel, connectionId) {
-    // do nothing by default
-    return {channel, connectionId};
-  }
-
 }
 
-module.exports = EntryPoint;
+module.exports = Manifest;
