+++ lib/api/core/entrypoints/protocols/http.js
@@ -22,9 +22,11 @@
 'use strict';
 
 const
-  errorsManager = require('../../../../../config/error-codes/throw').wrap('network', 'http'),
+  errorsManager = require('../../../../config/error-codes/throw').wrap(
+    'network',
+    'http'),
   assert = require('assert'),
-  debug = require('../../../../../kuzzleDebug')('kuzzle:entry-point:protocols:http'),
+  debug = require('../../../../kuzzleDebug')('kuzzle:entry-point:protocols:http'),
   bytes = require('bytes'),
   url = require('url'),
   ClientConnection = require('../clientConnection'),
@@ -49,6 +51,52 @@ const
     'Content-Encoding',
     'X-Kuzzle-Volatile'
   ].join(',');
+
+/**
+ * @class HttpContent
+ */
+class HttpMessage {
+  /**
+   * @param {ClientConnection} connection
+   * @param {http.IncomingMessage} request
+   */
+  constructor(connection, request) {
+    this.connection = connection;
+    this.raw = null;
+    this.json = null;
+    this.ips = connection.ips;
+    this.requestId = connection.id;
+    this.url = request.url;
+    this.method = request.method;
+    this.headers = request.headers;
+  }
+
+  addChunk (chunk) {
+    if (this.raw === null) {
+      this.raw = chunk;
+    }
+    else {
+      this.raw += chunk;
+    }
+  }
+
+  get content () {
+    if (this.json === null && this.raw !== null) {
+      try {
+        this.json = JSON.parse(this.raw);
+      }
+      catch (e) {
+        errorsManager.throw('unable_to_convert_http_body_to_json');
+      }
+    }
+
+    return this.json;
+  }
+
+  isEmpty () {
+    return this.json === null && this.raw === null;
+  }
+}
 
 /**
  * @class HttpProtocol
@@ -108,24 +156,16 @@ class HttpProtocol extends Protocol {
         `HTTP/${request.httpVersion}`,
         ips,
         request.headers),
-      // emulates a request coming from the (deprecated) Kuzzle Proxy server
-      proxyRequest = {
-        ips,
-        requestId: connection.id,
-        url: request.url,
-        method: request.method,
-        headers: request.headers,
-        content: ''
-      };
+      message = new HttpMessage(connection, request);
 
-    debug('[%s] receiving HTTP request: %a', connection.id, proxyRequest);
+    debug('[%s] Received HTTP request: %a', message);
+
     this.entryPoint.newConnection(connection);
 
     if (request.headers['content-length'] > this.maxRequestSize) {
       request.resume();
       this._replyWithError(
-        connection,
-        proxyRequest,
+        message,
         response,
         errorsManager.getError('http_request_size_exceeded'));
       return;
@@ -134,11 +174,11 @@ class HttpProtocol extends Protocol {
     let stream, pipes;
 
     try {
-      stream = this._createWritableStream(request, proxyRequest);
+      stream = this._createWritableStream(request, message);
       pipes = this._uncompress(request);
     } catch(err) {
       request.resume();
-      this._replyWithError(connection, proxyRequest, response, err);
+      this._replyWithError(message, response, err);
       return;
     }
 
@@ -163,13 +203,13 @@ class HttpProtocol extends Protocol {
       // registered pipes are NOT freed automatically
       pipes.forEach(pipe => pipe.close());
 
-      this._replyWithError(connection, proxyRequest, response, kerr);
+      this._replyWithError(message, response, kerr);
     });
 
     stream.on('finish', () => {
       debug('[%s] End Request', connection.id);
-      proxyRequest.headers['content-type'] = 'application/json';
-      this._sendRequest(connection, response, proxyRequest);
+      message.headers['content-type'] = 'application/json';
+      this._sendRequest(connection, response, message);
     });
   }
 
@@ -178,28 +218,24 @@ class HttpProtocol extends Protocol {
    *
    * @param {ClientConnection} connection
    * @param {http.ServerResponse} response
-   * @param {Object} proxyRequest
+   * @param {HttpMessage} message
    */
-  _sendRequest (connection, response, proxyRequest) {
-    debug('[%s] sendRequest: %a', connection.id, proxyRequest);
+  _sendRequest (connection, response, message) {
+    debug('[%s] sendRequest: %a', connection.id, message);
 
-    if (proxyRequest.json) {
-      proxyRequest.content = JSON.stringify(proxyRequest.json);
-    }
+    this.entryPoint.kuzzle.router.http.route(message, request => {
+      this.entryPoint.logAccess(request, message);
 
-    this.entryPoint.kuzzle.router.http.route(proxyRequest, request => {
-      this.entryPoint.logAccess(request, proxyRequest);
+      const data = this._getResponseData(request, message);
 
-      const data = this._getResponseData(request, proxyRequest);
-
-      this._compress(data, response, proxyRequest.headers, (err, deflated) => {
+      this._compress(data, response, message.headers, (err, deflated) => {
         if (err) {
-          const kuzerr = err instanceof KuzzleError 
+          const kuzerr = err instanceof KuzzleError
             ? err
             : errorsManager.getError('http_request_error', err.message);
-          this._replyWithError(connection, proxyRequest, response, kuzerr);
+          this._replyWithError(message, response, kuzerr);
           return;
-          
+
         }
 
         request.response.setHeader('Content-Length', String(deflated.length));
@@ -215,12 +251,11 @@ class HttpProtocol extends Protocol {
   /**
    * Forward an error response to the client
    *
-   * @param {ClientConnection} connection
-   * @param {Object} proxyRequest
+   * @param {HttpMessage} message
    * @param {http.ServerResponse} response
    * @param {Error} error
    */
-  _replyWithError(connection, proxyRequest, response, error) {
+  _replyWithError(message, response, error) {
     const result = {
       raw: true,
       content: JSON.stringify(removeErrorStack(error))
@@ -230,11 +265,11 @@ class HttpProtocol extends Protocol {
       ? error
       : errorsManager.getError('http_request_error', error);
 
-    debug('[%s] replyWithError: %a', connection.id, kerr);
+    debug('[%s] replyWithError: %a', message.connection.id, kerr);
 
     this.entryPoint.logAccess(
-      new Request(proxyRequest, {kerr, connectionId: connection.id}),
-      proxyRequest);
+      new Request(message, {error: kerr, connectionId: message.connection.id}),
+      message);
 
     response.writeHead(kerr.status, {
       'Content-Type': 'application/json',
@@ -246,7 +281,7 @@ class HttpProtocol extends Protocol {
 
     response.end(result.content);
 
-    this.entryPoint.removeConnection(connection.id);
+    this.entryPoint.removeConnection(message.connection.id);
   }
 
   /**
@@ -254,17 +289,17 @@ class HttpProtocol extends Protocol {
    * to send back to the client
    *
    * @param {Request} request
-   * @param {Object} proxyRequest
+   * @param {HttpMessage} message
    * @return {Buffer}
    */
-  _getResponseData(request, proxyRequest) {
+  _getResponseData(request, message) {
     let data = removeErrorStack(request.response.toJSON());
 
-    if (proxyRequest.requestId !== data.requestId) {
-      data.requestId = proxyRequest.requestId;
+    if (message.requestId !== data.requestId) {
+      data.requestId = message.requestId;
 
       if (!data.raw) {
-        data.content.requestId = proxyRequest.requestId;
+        data.content.requestId = message.requestId;
       }
     }
 
@@ -293,7 +328,7 @@ class HttpProtocol extends Protocol {
       }
     } else {
       let indent = 0;
-      const parsedUrl = url.parse(proxyRequest.url, true);
+      const parsedUrl = url.parse(message.url, true);
 
       if (parsedUrl.query && parsedUrl.query.pretty !== undefined) {
         indent = 2;
@@ -361,10 +396,10 @@ class HttpProtocol extends Protocol {
    * scratch configured to receive a HTTP request
    *
    * @param  {http.IncomingMessage} request
-   * @param  {Object} proxyRequest
+   * @param  {HttpMessage} message
    * @return {stream.Writable}
    */
-  _createWritableStream(request, proxyRequest) {
+  _createWritableStream(request, message) {
     if (request.headers['content-type'] &&
       !request.headers['content-type'].startsWith('application/json')
     ) {
@@ -374,7 +409,7 @@ class HttpProtocol extends Protocol {
             headers: request.headers,
             limits: { fileSize: this.maxFormFileSize }
           },
-          proxyRequest,
+          message,
           request);
       }
       catch (error) {
@@ -399,7 +434,7 @@ class HttpProtocol extends Protocol {
           return;
         }
 
-        proxyRequest.content += chunk.toString();
+        message.addChunk(chunk.toString());
         callback();
       }
     });
@@ -461,4 +496,4 @@ class HttpProtocol extends Protocol {
   }
 }
 
-module.exports = HttpProtocol;
+module.exports = { HttpMessage, HttpProtocol };
