+++ lib/api/core/entrypoints/protocols/mqtt.js
@@ -22,15 +22,15 @@
 'use strict';
 
 const
-  errorsManager = require('../../../../../config/error-codes/throw').wrap('network', 'mqtt'),
+  errorsManager = require('../../../../config/error-codes/throw').wrap(
+    'network',
+    'mqtt'),
   Bluebird = require('bluebird'),
-  debug = require('../../../../../kuzzleDebug')('kuzzle:entry-point:protocols:mqtt'),
+  debug = require('../../../../kuzzleDebug')('kuzzle:entry-point:protocols:mqtt'),
   ClientConnection = require('../clientConnection'),
   Protocol = require('./protocol'),
   mosca = require('mosca'),
-  {
-    Request
-  } = require('kuzzle-common-objects'),
+  { Request } = require('kuzzle-common-objects'),
   removeErrorStack = require('../removeErrorStack');
 
 /**
