+++ lib/api/core/entrypoints/protocols/socketio.js
@@ -23,14 +23,16 @@
 
 const
   _protocol = 'socketio',
-  debug = require('../../../../../kuzzleDebug')('kuzzle:entry-point:protocols:socketio'),
+  debug = require('../../../../kuzzleDebug')('kuzzle:entry-point:protocols:socketio'),
   ClientConnection = require('../clientConnection'),
   Protocol = require('./protocol'),
-  Request = require('kuzzle-common-objects').Request,
   {
-    KuzzleError,
-    InternalError: KuzzleInternalError
-  } = require('kuzzle-common-objects').errors;
+    Request,
+    errors: {
+      KuzzleError,
+      InternalError: KuzzleInternalError
+    }
+  } = require('kuzzle-common-objects');
 
 /**
  * @class SocketIoProtocol
