+++ lib/api/core/entrypoints/protocols/websocket.js
@@ -22,9 +22,11 @@
 'use strict';
 
 const
-  errorsManager = require('../../../../../config/error-codes/throw').wrap('network', 'websocket'),
+  errorsManager = require('../../../../config/error-codes/throw').wrap(
+    'network',
+    'websocket'),
   assert = require('assert'),
-  debug = require('../../../../../kuzzleDebug')('kuzzle:entry-point:protocols:websocket'),
+  debug = require('../../../../kuzzleDebug')('kuzzle:entry-point:protocols:websocket'),
   Protocol = require('./protocol'),
   { Server: WebSocketServer, Sender } = require('ws'),
   ClientConnection = require('../clientConnection'),
