+++ lib/api/core/entrypoints/removeErrorStack.js
