+++ lib/api/core/entrypoints/service/httpFormDataStream.js
@@ -22,19 +22,19 @@
 'use strict';
 
 const
-  SizeLimitError = require('kuzzle-common-objects').errors.SizeLimitError,
+  { errors: { SizeLimitError } } = require('kuzzle-common-objects'),
   Busboy = require('busboy');
 
 /**
  * @class HttpFormDataStream
  * @param {Object} opts
- * @param {Object} payload
+ * @param {HttpMessage} message
  * @param {http.IncomingMessage} request
  */
 class HttpFormDataStream extends Busboy {
-  constructor(opts, payload, request) {
+  constructor(opts, message, request) {
     super(opts);
-    payload.json = {};
+    message.json = {};
 
     this.on('file', (fieldname, file, filename, encoding, mimetype) => {
       let fileBuffer = Buffer.from('');
@@ -65,7 +65,7 @@ class HttpFormDataStream extends Busboy {
       });
 
       file.on('end', () => {
-        payload.json[fieldname] = {
+        message.json[fieldname] = {
           filename,
           encoding,
           mimetype,
@@ -76,7 +76,7 @@ class HttpFormDataStream extends Busboy {
     });
 
     this.on('field', (fieldname, val) => {
-      payload.json[fieldname] = val;
+      message.json[fieldname] = val;
     });
 
     return this;
