+++ lib/api/core/httpRouter/index.js
@@ -135,98 +135,102 @@ class Router {
   }
 
   /**
-   * Route an incoming HTTP httpRequest to the right handler
+   * Route an incoming HTTP message to the right handler
    *
-   * @param {object} httpRequest - HTTP httpRequest formatted by Kuzzle Proxy
+   * @param {HttpMessage} message - Parsed HTTP message
    * @param {function} cb
    */
-  route(httpRequest, cb) {
-    let request;
+  route(message, cb) {
+    debug('Routing HTTP message: %a', message);
 
-    debug('Routing request: %a', httpRequest);
-
-    if (!this.routes[httpRequest.method]) {
-      // the http protocol uses the requestId as connection id
-      const requestContext = this.kuzzle.router.connections.get(
-        httpRequest.requestId);
-
-      request = new Request(
-        { requestId: httpRequest.requestId },
-        requestContext && requestContext.toJSON());
-      request.response.setHeaders(this.defaultHeaders);
-
-      if (httpRequest.method.toUpperCase() === 'OPTIONS') {
-        request.input.headers = httpRequest.headers;
-        request.setResult({}, 200);
-
-        this.kuzzle.pipe('http:options', request)
-          .then(result => {
-            cb(result);
-            return null;
-          })
-          .catch(error => replyWithError(cb, request, error));
-
-        return;
-      }
-
-      replyWithError(
-        cb,
-        request,
-        errorsManager.getError('unrecognized_http_method', httpRequest.method));
-
-      return;
+    if (!this.routes[message.method]) {
+      return this.routeUnhandledHttpMethod(message, cb);
     }
 
-    httpRequest.url = httpRequest.url.replace(LeadingSlashRegex, '');
+    message.url = message.url.replace(LeadingSlashRegex, '');
+
+    let request;
 
     try {
-      const routeHandler = this.routes[httpRequest.method].getHandler(
-        httpRequest);
+      const routeHandler = this.routes[message.method].getHandler(message);
 
-      request = routeHandler.getRequest();
-      request.response.setHeaders(this.defaultHeaders);
+      routeHandler.request.response.setHeaders(this.defaultHeaders);
 
       if (routeHandler.handler === null) {
         errorsManager.throw('api_url_not_found', routeHandler.url);
       }
 
-      if (httpRequest.content.length <= 0) {
-        routeHandler.invokeHandler(cb);
-        return;
+      if (!message.isEmpty()) {
+        this.validateMessage(message);
       }
 
-      if (httpRequest.headers['content-type']
-      && !httpRequest.headers['content-type'].startsWith('application/json')) {
-        errorsManager.throw(
-          'invalid_request_content_type',
-          httpRequest.headers['content-type']);
-      }
-
-      {
-        const encoding = CharsetRegex.exec(httpRequest.headers['content-type']);
-
-        if (encoding !== null && encoding[1].toLowerCase() !== 'utf-8') {
-          errorsManager.throw(
-            'invalid_request_charset',
-            encoding[1].toLowerCase());
-        }
-      }
-
-      routeHandler.addContent(httpRequest.content);
       routeHandler.invokeHandler(cb);
-
-      return null;
     } catch (err) {
       if (request === undefined) {
-        request = new Request({requestId: httpRequest.requestId}, {}, 'rest');
+        request = new Request({requestId: message.requestId}, {});
         request.response.setHeaders(this.defaultHeaders);
       }
 
       const e = err instanceof KuzzleError
         ? err
         : errorsManager.getError('request_error', err);
+
       replyWithError(cb, request, e);
+    }
+  }
+
+  /**
+   * Route HTTP messages using an HTTP method that is not handled by Kuzzle's
+   * API, such as OPTIONS.
+   * @param  {HttpMessage} message
+   * @param {function} cb
+   */
+  routeUnhandledHttpMethod(message, cb) {
+    const
+      requestContext = this.kuzzle.router.connections.get(message.connection.id),
+      request = new Request(
+        { requestId: message.requestId },
+        requestContext && requestContext.toJSON());
+
+    request.response.setHeaders(this.defaultHeaders);
+
+    if (message.method.toUpperCase() === 'OPTIONS') {
+      request.input.headers = message.headers;
+      request.setResult({}, 200);
+
+      this.kuzzle.pipe('http:options', request)
+        .then(result => {
+          cb(result);
+          return null;
+        })
+        .catch(error => replyWithError(cb, request, error));
+
       return;
+    }
+
+    replyWithError(
+      cb,
+      request,
+      errorsManager.getError('unrecognized_http_method', message.method));
+
+    return;
+  }
+
+  /**
+   * Checks that an incoming HTTP message is well-formed
+   * @param  {HttpMessage} message
+   */
+  validateMessage (message) {
+    const contentType = message.headers['content-type'];
+
+    if ( contentType && !contentType.startsWith('application/json')) {
+      errorsManager.throw('invalid_request_content_type', contentType);
+    }
+
+    const encoding = CharsetRegex.exec(contentType);
+
+    if (encoding !== null && encoding[1].toLowerCase() !== 'utf-8') {
+      errorsManager.throw('invalid_request_charset', encoding[1].toLowerCase());
     }
   }
 }
