+++ lib/api/core/httpRouter/routeHandler.js
@@ -34,34 +34,45 @@ const
  * @class RouteHandler
  * @param {string} url - parsed URL
  * @param {object} query - parsed query string
- * @param {object} httpRequest - raw HTTP request informations
- * @throws {BadRequestError} If x-kuzzle-volatile HTTP header can not be parsed in JSON format
+ * @param {HttpMessage} message
+ * @throws {BadRequestError} If x-kuzzle-volatile HTTP header can not be parsed
+ *                           in JSON format
  */
 class RouteHandler {
-  constructor(url, query, httpRequest) {
+  constructor(url, query, message) {
     this.handler = null;
-    this.url = url;
-    this.request = null;
-    this.httpRequest = httpRequest;
-    this.data = {
-      requestId: this.httpRequest.requestId
+
+    const data = {
+      requestId: message.requestId,
+      body:message.content
     };
 
-    for (const k of Object.keys(this.httpRequest.headers)) {
-      if (k.toLowerCase() === 'authorization'
-        && this.httpRequest.headers[k].startsWith('Bearer ')
+    for (const k of Object.keys(message.headers)) {
+      if ( k.toLowerCase() === 'authorization'
+        && message.headers[k].startsWith('Bearer ')
       ) {
-        this.data.jwt = this.httpRequest.headers[k].substring('Bearer '.length);
-      } else if (k.toLowerCase() === 'x-kuzzle-volatile') {
+        data.jwt = message.headers[k].substring('Bearer '.length);
+      }
+      else if (k.toLowerCase() === 'x-kuzzle-volatile') {
         try {
-          this.data.volatile = JSON.parse(this.httpRequest.headers[k]);
+          data.volatile = JSON.parse(message.headers[k]);
         } catch (e) {
           errorsManager.throw('unable_to_convert_http_header_to_json');
         }
       }
     }
 
-    Object.assign(this.data, query);
+    this.request = new Request(
+      Object.assign(data, query),
+      {
+        connection: {
+          id: message.connection.id,
+          protocol: 'http',
+          url: message.url,
+          headers: message.headers,
+          ips: message.ips
+        }
+      });
   }
 
   /**
@@ -70,45 +81,7 @@ class RouteHandler {
    * @param {string} value
    */
   addArgument(name, value) {
-    this.data[name] = value;
-  }
-
-  /**
-   * Parse a string content and adds it to the right request object place
-   *
-   * @param {string} content
-   * @throws {BadRequestError} If the HTTP body can not be parsed in JSON format
-   */
-  addContent(content) {
-    try {
-      this.getRequest().input.body = JSON.parse(content);
-    } catch (e) {
-      errorsManager.throw('unable_to_convert_http_body_to_json');
-    }
-  }
-
-  /**
-   * Builds the request object and returns it
-   *
-   * @return {Request}
-   */
-  getRequest() {
-    if (this.request !== null) {
-      return this.request;
-    }
-
-    this.request = new Request(this.data, {
-      connection: {
-        id: this.data.requestId,
-        protocol: 'http',
-        url: this.url,
-        headers: this.httpRequest.headers,
-        ips: this.httpRequest.ips,
-        proxy: this.httpRequest.proxy
-      }
-    });
-
-    return this.request;
+    this.request.input.args[name] = value;
   }
 
   /**
@@ -117,7 +90,7 @@ class RouteHandler {
    * @param {Function} callback
    */
   invokeHandler(callback) {
-    this.handler(this.getRequest(), callback);
+    this.handler(this.request, callback);
   }
 }
