+++ lib/api/core/httpRouter/routePart.js
@@ -68,14 +68,14 @@ class RoutePart {
    * Returns a RouteHandler instance corresponding to the provided URL
    * Returns null if no handler was found
    *
-   * @param {object} request - HTTP request formatted by Kuzzle Proxy
+   * @param {HttpMessage} message
    * @return {RouteHandler} registered function handler
    */
-  getHandler(request) {
+  getHandler(message) {
     const
-      parsed = URL.parse(request.url, true),
+      parsed = URL.parse(message.url, true),
       pathname = parsed.pathname || '', // pathname is set to null if empty
-      routeHandler = new RouteHandler(pathname, parsed.query, request);
+      routeHandler = new RouteHandler(pathname, parsed.query, message);
 
     return getHandlerPart(this, pathname.split('/'), routeHandler);
   }
