+++ lib/api/core/janitor.js
@@ -181,10 +181,7 @@ class Janitor {
 
     return new Bluebird(resolve => {
       this.kuzzle.entryPoints.dispatch('shutdown');
-
-      // Gives time for the proxy to receive the shutdown signal
-      // and eventually to receive the last batch of requests to process
-      setTimeout(() => this._waitRemainingRequests(resolve), 1000);
+      this._waitRemainingRequests(resolve);
     });
   }
