+++ lib/api/core/notifier.js
@@ -26,11 +26,6 @@ const
   Request = require('kuzzle-common-objects').Request;
 
 /**
- * This internal service can either be invoked directly by
- * Kuzzle internal components, or through the internal
- * broker. This second way of communication is used
- * by workers to notify rooms about their work.
- *
  * @class NotifierController
  */
 class NotifierController {
