+++ lib/api/kuzzle.js
@@ -121,7 +121,7 @@ class Kuzzle extends EventEmitter {
     // Statistics core component
     this.statistics = new Statistics(this);
 
-    // Http server and websocket channel with server
+    // Network entry points
     this.entryPoints = new EntryPoints(this);
 
     // Validation core component
