+++ lib/config/index.js
@@ -41,15 +41,6 @@ function loadConfig () {
     }
   };
 
-  if (config.services.proxyBroker) {
-    // old configuration style. Disable embedded entry point
-    config.server.entryPoints.proxy = true;
-    config.server.entryPoints.embedded = false;
-
-    Object.assign(config.server.proxy, config.services.proxyBroker);
-    delete config.services.proxyBroker;
-  }
-
   // Injects the current Kuzzle version
   config.version = packageJson.version;
