+++ lib/services/index.js
@@ -23,7 +23,11 @@ const
   debug = require('debug')('kuzzle:services'),
   Bluebird = require('bluebird'),
   _ = require('lodash'),
-  KuzzleInternalError = require('kuzzle-common-objects').errors.InternalError;
+  {
+    errors: {
+      InternalError: KuzzleInternalError
+    }
+  } = require('kuzzle-common-objects');
 
 /**
  * @class Services
@@ -35,7 +39,6 @@ class Services {
      * @type {{
    *   internalCache: Redis,
    *   memoryStorage: Redis,
-   *   proxyBroker: ProxyBroker,
    *   storageEngine: ElasticSearch
    * }}
      */
@@ -45,61 +48,25 @@ class Services {
 
   /**
    * Initializes all services.
-   * Even if a service is in the blacklist option, it will be instantiated (it won't be initialized though).
-   * This allows togglable services.
-   * For instance, Kuzzle can be started with some services down by default, and toggled 'on' later.
-   *
-   * @param options may contains the following properties:
-   *            - blacklist (Array): the list of services that should not be initialized
    */
-  init(options) {
-    let
-      blacklist = [],
-      whitelist = null;
-
-    if (options) {
-      if (options.blacklist) {
-        blacklist = options.blacklist;
-      }
-      if (options.whitelist) {
-        whitelist = options.whitelist;
-        blacklist = [];
-      }
-    }
-
-    debug('initializing internal services:\nwhitelist: %O\nblacklist: %O', whitelist, blacklist);
-
+  init() {
     const promises = Object.keys(this.kuzzle.config.services)
-      .filter(key => {
-        // @deprecated - "internalBroker" was an internal service until Kuzzle 1.4.0
-        // Filtering this configuration allows users to upgrade their Kuzzle from pre-1.4.0
-        // without generating a breaking change because of an old configuration file
-        // This filter can be removed as soon as Kuzzle v1 support ends
-        if (key === 'internalBroker') {
-          this.kuzzle.log.warn('Ignoring the "internalBroker" service entry found in the kuzzlerc file. This service is deprecated and its corresponding entry in custom configuration files should be removed.');
-          return false;
-        }
-
-        return key !== 'common';
-      })
+      .filter(key => key !== 'common')
       .map(service => {
-        // We need to use a deferred promise here as the internalEngine (es) promises do not implement `finally`.
-        let opt = {service};
-
-        const init = whitelist
-          ? whitelist.indexOf(service) > -1
-          : blacklist.indexOf(service) === -1;
+        // We need to use a deferred promise here as the internalEngine (es)
+        // promises do not implement `finally`.
+        let opt = { service };
 
         return this.kuzzle.internalEngine
           .get('services', service)
           .then(response => {
             opt = _.merge(opt, response._source);
 
-            return registerService.call(this, service, opt, init);
+            return registerService.call(this, service, opt);
           })
           .catch(err => {
             if (err.status === 404) {
-              return registerService.call(this, service, opt, init);
+              return registerService.call(this, service, opt);
             }
 
             return Bluebird.reject(err);
@@ -114,30 +81,28 @@ class Services {
  * @this {Services}
  * @param serviceName
  * @param opts
- * @param init
  * @returns {Promise.<*>}
  */
-function registerService(serviceName, opts, init) {
+function registerService(serviceName, opts) {
   const
     services = this.kuzzle.config.services,
-    timeout = opts.timeout || this.kuzzle.config.services.common.defaultInitTimeout,
-    file = services[serviceName].backend
-      ? services[serviceName].backend
-      : serviceName,
-    aliases = services[serviceName].aliases
-      ? services[serviceName].aliases
-      : [serviceName];
+    timeout = opts.timeout || services.common.defaultInitTimeout,
+    file = services[serviceName].backend || serviceName,
+    aliases = services[serviceName].aliases || [serviceName];
 
   const promises = aliases.map(alias => {
     opts.service = alias;
 
-    debug('[%s] register internal service with alias %s from source %s', serviceName, alias, file);
+    debug(
+      '[%s] register internal service with alias %s from source %s',
+      serviceName,
+      alias,
+      file);
 
-    this.list[alias] = new (require('./' + file))(this.kuzzle, opts, services[serviceName]);
-
-    if (!init) {
-      return Bluebird.resolve();
-    }
+    this.list[alias] = new (require('./' + file))(
+      this.kuzzle,
+      opts,
+      services[serviceName]);
 
     return this.list[alias].init()
       .timeout(timeout, '___timedOut')
