+++ lib/util/shutdown.js
@@ -49,10 +49,7 @@ function runShutdown (kuzzle) {
 
   return new Bluebird(resolve => {
     kuzzle.entryPoints.dispatch('shutdown');
-
-    // Gives time for the proxy to receive the shutdown signal
-    // and eventually to receive the last batch of requests to process
-    setTimeout(() => waitRemainingRequests(kuzzle, resolve), 1000);
+    waitRemainingRequests(kuzzle, resolve);
   });
 }
