+++ test/mocks/kuzzle.mock.js
@@ -20,7 +20,6 @@ class KuzzleMock extends Kuzzle {
 
     // we need a deep copy here
     this.config = _.merge({}, config);
-    this.config.server.entryPoints.proxy = true;
 
     // emit + pipe mocks
     this.sandbox.stub(this, 'pipe').callsFake(
